nomad: fixup token policy validation
